fix(particle): native solver matches the GDScript oracle at the default lifetime
Three causes, each found by breaking it back in:
- quad height differences were taken in float32 (the oracle uses 64-bit);
- params were stored as float, so 0.05/0.3 arrived ~1e-9 off; now double,
  and the node rounds its own params through float32 so its GDScript route
  still equals the graph program's;
- the native solver ran droplets spawned on no-data or mask <= 0.001,
  which then cut the unmasked region; it now skips them like the oracle.
Grid accumulations round once on store, as the oracle's do.

Was 1.99 m max at 2000 droplets x 30 steps; now bitwise on all channels.
GraphHydraulicParticleGate [A3] default-lifetime parity and [A4] mask-spawn
parity, with measured-something controls and a completion count.

+++ src/pasture_3d_hydraulic_particle.h
@@ -16,17 +16,21 @@
 namespace godot {
 
 struct HydraulicParticleParams {
+	// DOUBLE, not float, for the reason ErosionHydraulicParams gives: the oracle's parameters are Variant
+	// doubles, and storing 0.05 or 0.3 as float made them arrive ~1e-9 off, which the droplets amplified
+	// into a metre by the default lifetime (GraphHydraulicParticleGate [A3]). The graph path's float32
+	// program params widen to the same number, so it is unaffected.
 	int droplet_count = 25000;
 	int max_lifetime = 30;
-	float inertia = 0.05f;
-	float sediment_capacity = 4.0f;
-	float erosion_speed = 0.3f;
-	float deposition_speed = 0.3f;
-	float evaporation_rate = 0.01f;
-	float min_slope = 0.01f;
-	float gravity = 4.0f;
-	float bedrock_gap = 2.0f;
-	float ridge_forcing = 0.0f;
+	double inertia = 0.05;
+	double sediment_capacity = 4.0;
+	double erosion_speed = 0.3;
+	double deposition_speed = 0.3;
+	double evaporation_rate = 0.01;
+	double min_slope = 0.01;
+	double gravity = 4.0;
+	double bedrock_gap = 2.0;
+	double ridge_forcing = 0.0;
 	int64_t seed = 1337;
 	PackedFloat32Array mask;
